% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rscodeio_diagnose}
\alias{rscodeio_diagnose}
\title{Get diagnostic information about the rscodeiov2 installation}
\usage{
rscodeio_diagnose()
}
\value{
A list containing diagnostic information
}
\description{
Get diagnostic information about the rscodeiov2 installation
}